Frontend edits - delete all folder code An absolutely massive commit because I'm insane and randomly change things that don't need to be! No API changes yet. Also reorganizes some parts of the app, namely moving some utils/ modules to their better place in app/(ui)/.
