+++ app/(ui)/tempresourcedelete.module.tsx
@@ -3,13 +3,13 @@ import { faFileCircleMinus } from "@fortawesome/free-solid-svg-icons";
 import { FontAwesomeIcon } from "@fortawesome/react-fontawesome";
 import { useRouter } from "next/navigation";
 
-export function DeleteResource({ projectID, resourceID, folderID }: { projectID: string, resourceID: string, folderID?: string }) {
+export function DeleteResource({ projectID, resourceID }: { projectID: string, resourceID: string }) {
   const router = useRouter();
-  function onClick(projectID: string, resourceID: string, folderID?: string) {
+  function onClick(projectID: string, resourceID: string) {
     if (!confirm("Are you sure you want to delete this resource?")) return;
     fetch(`/api/resource`, {
       method: "DELETE",
-      body: JSON.stringify({ projectID, resourceID, folderID }),
+      body: JSON.stringify({ projectID, resourceID }),
     })
       .then((res) => {
         return res.json();
@@ -22,5 +22,5 @@ export function DeleteResource({ projectID, resourceID, folderID }: { projectID:
         }
       });
   }
-  return <button className="hover:text-sky-500" onClick={() => onClick(projectID, resourceID, folderID)}><FontAwesomeIcon icon={faFileCircleMinus} /></button>
+  return <button className="hover:text-sky-500" onClick={() => onClick(projectID, resourceID)}><FontAwesomeIcon icon={faFileCircleMinus} /></button>
 }
