+++ app/project/[id]/new/page.tsx
@@ -1,6 +1,6 @@
 import { auth } from "@/auth";
 import { getProject } from "@/utils/db"
-import FormattedDate from "@/utils/time.module";
+import FormattedDate from "@/app/(ui)/time.module";
 import { NotFoundProject } from "@/app/(ui)/notfound.module";
 import FullCreateResource from "./createresource.module";
 import { FontAwesomeIcon } from "@fortawesome/react-fontawesome";
