+++ app/project/[id]/view/[resourceid]/resourcetools.module.tsx
@@ -4,7 +4,7 @@ import { FontAwesomeIcon } from "@fortawesome/react-fontawesome";
 import { useState } from "react";
 import { useRouter } from "next/navigation";
 
-export function ResourceTools({ projectID, resourceID, folderID }: { projectID: string, resourceID: string, folderID?: string }) {
+export function ResourceTools({ projectID, resourceID }: { projectID: string, resourceID: string }) {
   type PossibleMenus = "deleteresource" | null;
   const [openMenu, setOpenMenu] = useState<PossibleMenus>(null);
   return (
@@ -12,12 +12,12 @@ export function ResourceTools({ projectID, resourceID, folderID }: { projectID: 
       <div className="flex gap-4 bg-slate-200 dark:bg-slate-800 p-4 rounded-lg justify-between sticky">
         <button onClick={() => setOpenMenu(openMenu !== "deleteresource" ? "deleteresource" : null)} className="hover:text-sky-500"><FontAwesomeIcon icon={faTrashCan} /></button>
       </div>
-      {openMenu === "deleteresource" && <DeleteResource projectID={projectID} resourceID={resourceID} closeMenu={() => setOpenMenu(null)} folderID={folderID} />}
+      {openMenu === "deleteresource" && <DeleteResource projectID={projectID} resourceID={resourceID} closeMenu={() => setOpenMenu(null)} />}
     </div>
   )
 }
 
-function DeleteResource({ projectID, resourceID, closeMenu, folderID }: { projectID: string, resourceID: string, closeMenu: () => void, folderID?: string }) {
+function DeleteResource({ projectID, resourceID, closeMenu }: { projectID: string, resourceID: string, closeMenu: () => void }) {
   const router = useRouter();
   function onClick(projectID: string, resourceID: string, folderID?: string) {
     fetch(`/api/resource`, {
@@ -38,7 +38,7 @@ function DeleteResource({ projectID, resourceID, closeMenu, folderID }: { projec
   return (
     <div className="flex flex-col gap-4 bg-slate-200 dark:bg-slate-800 p-4 rounded-lg my-2">
       <p className="text-red-500">Are you sure you want to delete this resource?</p>
-      <button onClick={() => onClick(projectID, resourceID, folderID)} className="bg-red-500 p-2 rounded-xl border-2 border-slate-400 dark:border-slate-900 hover:text-sky-500"><FontAwesomeIcon icon={faTrash} /> Delete Resource</button>
+      <button onClick={() => onClick(projectID, resourceID)} className="bg-red-500 p-2 rounded-xl border-2 border-slate-400 dark:border-slate-900 hover:text-sky-500"><FontAwesomeIcon icon={faTrash} /> Delete Resource</button>
       <button onClick={() => closeMenu()} className="hover:text-sky-500 bg-slate-300 dark:bg-slate-700 p-2 text-center rounded-xl border-2 border-slate-400 dark:border-slate-900">Cancel</button>
     </div>
   )
